fix: 🚨 Somes fix linter error

+++ src/utilities/db/Birthday.ts
@@ -3,14 +3,6 @@ import type { Dayjs } from 'dayjs';
 import type { User } from 'discord.js';
 
 export class Birthday extends Utility {
-	private prisma = this.container.prisma;
-
-	public constructor(context: Utility.Context, options: Utility.Options) {
-		super(context, {
-			...options,
-			name: 'birthday',
-		});
-	}
 	public get = {
 		BirthdaysByDate: (date: Dayjs) => this.prisma.birthday.findMany({ where: { birthday: { contains: date.format('-MM-DD') } } }),
 		BirthdayByDateAndTimezone: (date: Dayjs, timezone: number) =>
@@ -18,54 +10,62 @@ export class Birthday extends Utility {
 		BirthdaysByGuildId: (guildId: string) => this.prisma.birthday.findMany({ where: { guildId } }),
 		BirthdayByUserAndGuild: (guildId: string, userId: string) =>
 			this.prisma.birthday.findUnique({
-				where: { userId_guildId: { guildId, userId } },
+				where: { userId_guildId: { guildId, userId } }
 			}),
-		BirthdaysNotDisabled: (guildId: string) => this.prisma.birthday.findMany({ where: { guildId, disabled: false } }),
+		BirthdaysNotDisabled: (guildId: string) => this.prisma.birthday.findMany({ where: { guildId, disabled: false } })
 	};
 
 	public update = {
 		BirthdayDisabled: (guildId: string, userId: string, disabled: boolean) =>
 			this.prisma.birthday.update({
 				where: { userId_guildId: { guildId, userId } },
-				data: { disabled },
+				data: { disabled }
 			}),
 		BirthdayByUserAndGuild: (guildId: string, userId: string, birthday: string) =>
 			this.prisma.birthday.update({
 				where: { userId_guildId: { guildId, userId } },
-				data: { birthday },
-			}),
+				data: { birthday }
+			})
 	};
 
 	public delete = {
 		GuildById: (guildId: string) => this.prisma.guild.delete({ where: { guildId } }),
 		ByDisabledGuilds: () => this.prisma.guild.deleteMany({ where: { disabled: true } }),
 		ByLastUpdatedDisabled: (date: Date) => this.prisma.guild.deleteMany({ where: { lastUpdated: { lt: date.toISOString() }, disabled: true } }),
-		ByGuildAndUser: (guildId: string, userId: string) =>
-			this.prisma.birthday.delete({ where: { userId_guildId: { guildId, userId } } }),
+		ByGuildAndUser: (guildId: string, userId: string) => this.prisma.birthday.delete({ where: { userId_guildId: { guildId, userId } } })
 	};
+
+	private prisma = this.container.prisma;
+
+	public constructor(context: Utility.Context, options: Utility.Options) {
+		super(context, {
+			...options,
+			name: 'birthday'
+		});
+	}
 
 	public create = (birthday: string, guildId: string, user: User) =>
 		this.prisma.birthday.create({
 			data: {
-				birthday: birthday,
+				birthday,
 				guild: {
 					connectOrCreate: {
 						create: { guildId },
-						where: { guildId },
-					},
+						where: { guildId }
+					}
 				},
 				user: {
 					connectOrCreate: {
 						create: {
 							userId: user.id,
 							discriminator: user.discriminator,
-							username: user.username,
+							username: user.username
 						},
 						where: {
-							userId: user.id,
-						},
-					},
-				},
-			},
+							userId: user.id
+						}
+					}
+				}
+			}
 		});
 }
